nuclei-mcp/requirements.txt
requests
pyyaml
python-dotenv
beautifulsoup4
flask